Used to interface with giga to control motors manually via bluetooth
This code uses the Bluepad Library to interface with any Bluetooth controller/device. This allows for manual control of the car robot with a ps4 controller due to incompatibilities between current Bluetooth libraries and the Arduino Giga. The method of communication between Nano and Giga in this use is the I2C Bus with the Nano functioning as the slave device.